feat: Inlining, updating and improving sns geocountry lookup
To identify the aws endpoint for sns lookups the tradle/countries repository was previously used. This repository however is lacking some modern aws repositories and countries and was not published to npm. Also two other points were not respected. For one: not all aws endpoints support sns before this commit a fallback would be chosen by closest "parent" group. Now it uses the geograpically closest, working endpoint. In a minor point: the AWS documentation also specifies the countries to which a sns call is possibe. This information is now included in the code.