chore: adding clean script to make sure its a clean compilation
When typescript compiles files, it doesn't delete other files in the target directory. This means when files are deleted in the source directory they may still stick around in the target directory. Furthermore other folders (like node_modules) might interfere in the build process. With a clean script run before the publication we can make sure that no extraneous files get accidentally published and that its is run against the latest packages. Furthermore it is important to note that bootstrap will create node_modules folders in the sub-packages for .bin files that will interfere with the module resolution. tsbuildinfo needs to be deleted also as typescript simply doesn’t build files that havn’t changed.

Note: It took quite a long time to figure out these three things (-_-)