feat: forwarding fields of @tradle/aws-info instead of providing the data directly.
The @tradle/aws-common-utils package provides **functionality** which doesn't need to be update if the data updates. For that reason all constants are now held in @tradle/aws-info. Using the loose dependency definition of "^1.0.0" makes it possible for mycloud to update aws-info without the need for updating the other @tradle/aws-* packages. This is intended to reduce the maintenance necessary for updating the aws-info packages. It is expected that aws-info respects the semver as far as data-structure is concerned but that the content may change.