improvements: aws-info refactor and extended (see below)
This combines several strategies to reduce the startup load for using aws-info. For one it loads(requires) data only on-demand which means that simply referencing aws-info does not cause loading it. Also it reduces the number of files which factors into load time. As the data is now loaded on demand the JS objects are now not strictly constants anymore, which causes the naming change from constant to instance naming.

On the additional features added is now partitioning which defines partitions for all aws regions and geo-locations for us-gov cloud (assumed at the same locations as the us-east/west clouds, as location seems to not be public knowledge). When looking for fallback region (ses/sns) it will only look within the same partition.

It also adds the "lognestCallingCode" constant that was previously computed on startup.