fix: changing way to establish region for sms dispatching.
The previous algorithm looked for the geographically nearest region to the telephone number for dispatching a sms call. We decided to change the algorithm to use the current region, if sms sending is supported. Without support it will use the .fallback information of the region (compiled in @tradle/aws-info) that contains the geographically nearest region that does support sms sending. This way the information of the telephone number doesn't leave the region unless it has to.

Note: At this point in time there is no region that we support that does not support sms, except for the entire china partition. However: in future, new regions may not support this, which is why we keep this logic.